--- a/src/sys/timer.cxx
+++ b/src/sys/timer.cxx
@@ -4,6 +4,7 @@ Timer::Timer() : timing(getInfo("")) {
   if (timing.counter == 0) {
     timing.started = clock_type::now();
     timing.running = true;
+    timing.ntimes++;
   }
   timing.counter += 1;
 }
@@ -12,6 +13,7 @@ Timer::Timer(const std::string& label) : timing(getInfo(label)) {
   if (timing.counter == 0) {
     timing.started = clock_type::now();
     timing.running = true;
+    timing.ntimes++;
   }
   timing.counter += 1;
 }
@@ -33,7 +35,7 @@ Timer::timer_info& Timer::getInfo(const std::string& label) {
   auto it = info.find(label);
   if (it == info.end()) {
     auto timer = info.emplace(
-      label, timer_info{seconds{0}, false, clock_type::now(), 0});
+                              label, timer_info{seconds{0}, false, clock_type::now(), 0, 0});
     return timer.first->second;
   }
   return it->second;