--- a/src/sys/timer.cxx
+++ b/src/sys/timer.cxx
@@ -74,40 +74,50 @@ double Timer::resetTime(Timer::timer_info& info) {
 }
 
 void Timer::listAllInfo() {
-  const std::string headerName = "Timer name";
-  const std::string headerTime = "Current time (s)";
-  const std::string headerTotal = "Total time (s)";
-  const std::string headerHits = "Hits";
+  using namespace std::string_literals;
+  const auto header_name = "Timer name"s;
+  const auto header_time = "Total time (s)"s;
+  const auto header_hits = "Hits"s;
+  const auto header_mean = "Mean time/hit (s)"s;
 
+  // Helper lambda to get the longest item in info, after applying
+  // another lambda (op), and also taking into account the header. op
+  // transforms from a map key-value pair to the length of either the
+  // key or a member of the value (Timer::timing_info). As
+  // std::max_element returns an iterator, it needs to be dereferenced
+  // and have op applied to it
   const auto longest_width = [](const std::string& header, const auto& op) {
-    return std::max(header.size(),
+    return std::max(header.length(),
                     op(*std::max_element(std::begin(info), std::end(info),
                                          [&op](const auto& lhs, const auto& rhs) {
                                            return op(lhs) < op(rhs);
                                          })));
   };
 
+  // Get the widths of the four columns: timer name, total time,
+  // number of hits, mean time per hit
   const auto max_width =
-      longest_width(headerName, [](const auto& it) { return it.first.length(); });
-  const auto time_width = longest_width(headerTime, [](const auto& it) {
-    return fmt::format("{:.9g}", it.second.time.count()).length();
-  });
-  const auto total_time_width = longest_width(headerTotal, [](const auto& it) {
+      longest_width(header_name, [](const auto& it) { return it.first.length(); });
+  const auto total_time_width = longest_width(header_time, [](const auto& it) {
     return fmt::format("{:.9g}", it.second.total_time.count()).length();
   });
-  const auto hit_width = longest_width(headerHits, [](const auto& it) {
+  const auto hit_width = longest_width(header_hits, [](const auto& it) {
     return fmt::format("{}", it.second.hits).length();
   });
+  // Assume that the mean will be about the same length as the total
+  const auto mean_time_width = longest_width(header_mean, [](const auto& it) {
+    return fmt::format("{:.9g}", it.second.total_time.count()).length();
+  });
 
-  using namespace std::string_literals;
-
-  output.write("{0:<{1}} | {2:<{3}} | {5:<{6}} | {4}\n", headerName, max_width,
-               headerTime, time_width, headerHits, headerTotal, total_time_width);
-  output.write("{0:-<{1}}-|-{0:-<{2}}-|-{0:-<{4}}-|-{0:-<{3}}\n", ""s, max_width,
-               time_width, hit_width, total_time_width);
+  output.write("{0:<{1}} | {2:<{3}} | {4:<{5}} | {6:<{7}}\n", header_name, max_width,
+               header_time, total_time_width, header_hits, hit_width, header_mean,
+               mean_time_width);
+  output.write("{0:-<{1}}-|-{0:-<{2}}-|-{0:-<{3}}-|-{0:-<{4}}\n", ""s, max_width,
+               total_time_width, hit_width, mean_time_width);
   for (const auto& kv : info) {
-    output.write("{0:<{1}} | {2:<{3}.9g} | {6:<{7}.9g} | {4:<{5}}\n", kv.first, max_width,
-                 kv.second.time.count(), time_width, kv.second.hits, hit_width,
-                 kv.second.total_time.count(), total_time_width);
+    output.write("{0:<{1}} | {2:<{3}.9g} | {4:<{5}} | {6:<{7}.9g}\n", kv.first, max_width,
+                 kv.second.total_time.count(), total_time_width, kv.second.hits,
+                 hit_width, kv.second.total_time.count() / kv.second.hits,
+                 mean_time_width);
   }
 }