--- a/src/sys/timer.cxx
+++ b/src/sys/timer.cxx
@@ -77,6 +77,7 @@ void Timer::listAllInfo() {
   using namespace std::string_literals;
   const auto header_name = "Timer name"s;
   const auto header_time = "Total time (s)"s;
+  const auto header_frac = "% of top"s;
   const auto header_hits = "Hits"s;
   const auto header_mean = "Mean time/hit (s)"s;
 
@@ -101,6 +102,8 @@ void Timer::listAllInfo() {
   const auto total_time_width = longest_width(header_time, [](const auto& it) {
     return fmt::format("{:.9g}", it.second.total_time.count()).length();
   });
+  // Only print percentages to two d.p.
+  const auto frac_width = std::max(header_frac.length(), "100.00%"s.length());
   const auto hit_width = longest_width(header_hits, [](const auto& it) {
     return fmt::format("{}", it.second.hits).length();
   });
@@ -109,15 +112,29 @@ void Timer::listAllInfo() {
     return fmt::format("{:.9g}", it.second.total_time.count()).length();
   });
 
-  output.write("{0:<{1}} | {2:<{3}} | {4:<{5}} | {6:<{7}}\n", header_name, max_width,
-               header_time, total_time_width, header_hits, hit_width, header_mean,
-               mean_time_width);
-  output.write("{0:-<{1}}-|-{0:-<{2}}-|-{0:-<{3}}-|-{0:-<{4}}\n", ""s, max_width,
-               total_time_width, hit_width, mean_time_width);
-  for (const auto& kv : info) {
-    output.write("{0:<{1}} | {2:<{3}.9g} | {4:<{5}} | {6:<{7}.9g}\n", kv.first, max_width,
-                 kv.second.total_time.count(), total_time_width, kv.second.hits,
-                 hit_width, kv.second.total_time.count() / kv.second.hits,
-                 mean_time_width);
+  output.write("{0:<{1}} | {2:<{3}} | {4:<{5}} | {6:<{7}} | {8:<{9}}\n", header_name,
+               max_width, header_time, total_time_width, header_frac, frac_width,
+               header_hits, hit_width, header_mean, mean_time_width);
+  output.write("{0:-<{1}}-|-{0:-<{2}}-|-{0:-<{3}}-|-{0:-<{4}}-|-{0:-<{5}}\n", ""s,
+               max_width, total_time_width, frac_width, hit_width, mean_time_width);
+
+  // To sort the map by the total time, we need to copy the key-value
+  // pairs into another container that we can sort ourselves
+  std::vector<std::pair<std::string, timer_info>> sorted_info;
+  sorted_info.reserve(info.size());
+  std::transform(begin(info), end(info), std::back_inserter(sorted_info),
+                 [](const auto& it) { return it; });
+  // Sort so that the largest total time is first
+  std::sort(begin(sorted_info), end(sorted_info), [](const auto& a, const auto& b) {
+    return a.second.total_time > b.second.total_time;
+  });
+
+  for (const auto& kv : sorted_info) {
+    output.write("{0:<{1}} | {2:<{3}.9g} | {8:<{9}.2%} | {4:<{5}} | {6:<{7}.9g}\n",
+                 kv.first, max_width, kv.second.total_time.count(), total_time_width,
+                 kv.second.hits, hit_width, kv.second.total_time.count() / kv.second.hits,
+                 mean_time_width,
+                 kv.second.total_time.count() / sorted_info[0].second.total_time.count(),
+                 frac_width);
   }
 }