--- a/tests/unit/sys/test_timer.cxx
+++ b/tests/unit/sys/test_timer.cxx
@@ -254,6 +254,6 @@ TEST(TimerTest, ListAllInfo) {
 
   using namespace ::testing;
   EXPECT_THAT(cout_capture.str(), HasSubstr("Timer name         |"));
-  EXPECT_THAT(cout_capture.str(), ContainsRegex("one *| 0\\.\\d+ | 0\\.\\d+ | 1"));
-  EXPECT_THAT(cout_capture.str(), ContainsRegex("two *| 0 *| 0\\.\\d+ | 2"));
+  EXPECT_THAT(cout_capture.str(), ContainsRegex("one *| 0\\.\\d+ | 1    | 0\\.\\d+"));
+  EXPECT_THAT(cout_capture.str(), ContainsRegex("two *| 0 * | 2    | 0\\.\\d+"));
 }