Adds two methods to `Timer`:

- `getTotalTime` to see the total time elapsed for a timer (i.e. doesn't get reset with `resetTime`)
- `listAllInfo` generates a table reporting all the timers, like:

```
Timer name | Current time (s) | Total time (s) | Hits
-----------|------------------|----------------|-----
comms      | 0                | 0.028078329    | 546 
evolve_rho | 0.499829339      | 0.499829339    | 62  
invert     | 0                | 0.539184925    | 62  
io         | 0.010658816      | 0.090785536    | 57  
rhs        | 0                | 4.15958722     | 62  
run        | 0.014304425      | 4.32613615     | 1   
```

Adds an option `write_final_timings` which prints the above table at the end of a simulation.

I'm not so sure the "current time" column is particularly useful, probably a "mean time per hit" would be better. Also I'm not overly keen on the name `write_final_timings` at global scope, so suggestions welcome!