@ZedThree 2020-01-09
I've changed the columns to drop the current time and add mean time per hit:

```
Timer name | Total time (s) | Hits | Mean time/hit (s)
-----------|----------------|------|------------------
comms      | 0.019139448    | 546  | 3.50539341e-05   
evolve_rho | 0.357001864    | 62   | 0.00575809458    
invert     | 0.376253795    | 62   | 0.0060686096     
io         | 0.055955419    | 57   | 0.000981674018   
rhs        | 2.97283263     | 62   | 0.0479489133     
run        | 3.07722771     | 1    | 3.07722771   
```

Maybe the switch should be `show_timing_report`?