@bendudson 2020-01-10
Would sorting by one of the columns be simple to do? e.g. sort by total time might be the most useful most of the time