@ZedThree 2020-01-17
I've changed the option to `time_report:show`, sticking it in its own section for a couple of reasons: I'd prefer not to add stuff to the global section if possible, and to not preclude future customisation options.

Also added a column with fraction of the largest time. Now looks like:

```
Timer name | Total time (s) | % of top | Hits | Mean time/hit (s)
-----------|----------------|----------|------|------------------
run        | 4.37208176     | 100.00%  | 1    | 4.37208176       
rhs        | 4.18987402     | 95.83%   | 62   | 0.0675786133     
invert     | 0.543508674    | 12.43%   | 62   | 0.00876626894    
evolve_rho | 0.504323351    | 11.54%   | 62   | 0.0081342476     
io         | 0.101402551    | 2.32%    | 57   | 0.00177899212    
comms      | 0.029098885    | 0.67%    | 546  | 5.32946612e-05   
```